+++ app/src/main/java/com/emikhalets/mydates/ui/calendar/CalendarVM.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.emikhalets.mydates.data.database.ListResult
 import com.emikhalets.mydates.data.repositories.DatabaseRepository
-import com.emikhalets.mydates.utils.dayOfYear
+import com.emikhalets.mydates.utils.extentions.dayOfYear
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
