+++ app/src/main/java/com/emikhalets/mydates/ui/calendar/DayEventsAdapter.kt
@@ -10,7 +10,7 @@ import com.emikhalets.mydates.R
 import com.emikhalets.mydates.data.database.entities.Event
 import com.emikhalets.mydates.databinding.ItemEventBinding
 import com.emikhalets.mydates.utils.enums.EventType
-import com.emikhalets.mydates.utils.formatDate
+import com.emikhalets.mydates.utils.extentions.formatDate
 
 class DayEventsAdapter(
     private val click: (Event) -> Unit
