+++ app/src/main/java/com/emikhalets/mydates/utils/enums/EventType.kt
@@ -2,7 +2,7 @@ package com.emikhalets.mydates.utils.enums
 
 import android.content.Context
 import com.emikhalets.mydates.R
-import com.emikhalets.mydates.utils.formatDate
+import com.emikhalets.mydates.utils.extentions.formatDate
 
 enum class EventType(val value: Int) {
     ANNIVERSARY(1),
