+++ app/src/main/java/com/emikhalets/mydates/utils/extentions/ViewsExtensions.kt
@@ -1,6 +1,5 @@
-package com.emikhalets.mydates.utils
+package com.emikhalets.mydates.utils.extentions
 
-import android.view.View
 import android.widget.EditText
 import android.widget.TextView
 import androidx.annotation.DrawableRes
@@ -13,12 +12,7 @@ fun TextView.setDrawableStart(@DrawableRes resource: Int) {
     this.setCompoundDrawablesRelativeWithIntrinsicBounds(resource, 0, 0, 0)
 }
 
-fun EditText.setDate(date: Long, withoutYear: Boolean) {
+fun EditText.setDateText(date: Long, withoutYear: Boolean) {
     if (withoutYear) this.setText(date.formatDate("d MMMM"))
     else this.setText(date.formatDate("d MMMM y"))
-}
-
-fun View.setVisibility(visibility: Boolean) {
-    if (visibility) this.visibility = View.VISIBLE
-    else this.visibility = View.GONE
 }
